pyproject.toml
[project]
name = "educhain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "fastapi-users[sqlachemy]>=15.0.3",
    "pwdlib[bcrypt]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
